requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3
motor==3.3.2
pymongo==4.6.1
cloudinary==1.37.0
PyPDF2==3.0.1
google-generativeai==0.3.2